tapdb: move all cache related code to own file
